libsyn123: more smoothness on rate ratio changes
This introduces the relaxing of filters using a bit of historic
samples at the differing decimation stages. New stages get
initialized using the previous' stages history. Final pre-emphasis
and lowpass, as well as interpolator histoy get fed the (constructed)
final stage history to get on track. This does work nicely to remove
(dampen, of course) the clicks that were still audible with only the
low pass coefficient scaling.

The sweeper test still shows some clicking towards the end of a
speed-up run, but that maybe is due to the output sample interval
getting smaller and smaller, not really giving space to smoothness.
I need a smarter sweeper to really compare a resampled sweep to
a computed sine sweep.

Currently, there is an additional parameter to syn123_setup_resample()
to activate the not-for-free stage history keeping. I will do some
benchmarking before deciding whether to keep that parameter or just
have the full smoothness always present.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4522 35dc7657-300d-0410-a2e5-dc2837fedb53