build system: fix up library dependencies, add Libs.private to .pc
This separates the dependencies of our libraries from the dependencies
of the end-user programs. For this, libcompat got split again, to
separate the libdl-using part for libout123 only. Messy, but seems to
work.

The pkg-config files now contain Libs.private to enable proper usage
of our static libs. This is mainly interesting for Windows, but also
covers libm usage on Unix.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4551 35dc7657-300d-0410-a2e5-dc2837fedb53