mpg123: filter metadata before printout, better handling of Unicode width
This is prompted by bug 267. The main goal was to avoid control characters
in terminal printout. Another result is the more strict handling of ID3v1
data, it being reduced to 7 bits in lack of a safe alternative.

In the best case, proper handling of UTF-8 strings should be ensured using
the system Unicode support (mbstowcs(), iswprint() and friends). Fallback
code is in place to do something sensible if that support is missing. With
the system support, widths of characters different from 1 are handled now,
as correctly as the evolving Unicode standard allows.

This has been tested on Linux only. At least on Windows, there will be
additional handling needed, I imagine configuring the terminal to acceppt
UTF-8 at all is in order.

Also, the filtering needs to be extended to file names and URLs.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4570 35dc7657-300d-0410-a2e5-dc2837fedb53