libout123/alsa: set buffer_size/2 as start threshold
This seems to fix the half-broken playback with pulse hidden behind ALSA
api via a plugin. At least on ubuntu, I can avoid initial underruns that
were accompanied by some periods of static noise. I presume that premature
starting of the playback is also to blame for the continued static noise
that some people get from mpg123 with ALSA output module routing to a
Pulseaudio server.

I doubt that this is the actual cause for the messup that seems to run
into some byte shift shredding the decoded audio, but it's a trigger.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4572 35dc7657-300d-0410-a2e5-dc2837fedb53