libmpg123: ID3v2 cleanup 1 (properly append tags with overwriting)
This implements proper behaviour for the assumption that multiple encountered
ID3v2 tags shall be applied as updates, adding only to the stored elements
if existing entries should not be overwritten according to the uniqueness
rules of the standard.

An easy way to test this is to decode a track with --seek-index 0 and
repeatedly seek back to the beginning (and add a line to print out
the number of collected ID3 texts/comments). The old libmpg123 would
steadily append the same entries again and again to the id3 data
structures as the leading ID3v2 tag is parsed repeatedly. This is
strictly speaking no memory leak as the pointers are still properly
managed, but of course it is a leak in practice as data accumulates
needlessly.

Note that this accumulation only happened when the user explicitly
disabled the seek index for regular files, which would prevent repeated
parsing of a leading ID3v2 tag, or in the case of concatenated files
in one big stream, where libmpg123 would remember all ID3v2 data
ever encountered. It is not easy to judge the bugginess of this behaviour.
It could actually be a feature to accumulate all ID3 data. And users
could call mpg123_meta_free() after each extraction to minimize memory
usage. Nevertheless, this ends now in the sense that ID3v2 frames/fields
that are supposed to be unique are also unique in storage.

The next step will be to add parsing of the update flag and clearing
of existing ID3v2 data when a non-update tag is parsed.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4608 35dc7657-300d-0410-a2e5-dc2837fedb53