mpg123: finish up the printout filtering (bug 267)
This now filters strings from libmpg123 metadata and the user/web input for useless/nasty
control/non-printable characters when the output is a terminal. It covers all relevant
printouts except the ones I forgot. Metadata, user-provided command-line arguments and
file names in the playlist are covered. I guess really relevant is the data that the
user did not enter explicitly (metadata in streams, playlist from URL).

The proper way would be to fully use locales and a curses library to sanitize terminal
output, but I really would like to avoid that mess, having my personal mess as a compromise
that should mostly work OK. Future may correct me. When we want to do it properly, we should
re-write the whole mpg123 application from scratch, using the mpg123 libraries for audio and
other libraries for the interface.

git-svn-id: svn://scm.orgis.org/mpg123/trunk@4636 35dc7657-300d-0410-a2e5-dc2837fedb53